requirements.txt
numpy~=1.20.2
pillow~=8.2.0
opencv-python~=4.5.1.48
PyDirectInput~=1.0.4
matplotlib~=3.4.1